- Fixed reading packets(Will read 100% of the packet.) - Supports all post request types (as far as I know) - Minor change to the TcpClient wrapped class.

+++ BackedFramework/Resources/Extensions/TcpClient.cs
@@ -47,12 +47,12 @@ namespace BackedFramework.Resources.Extensions
             }
         }
 
-        public void ReadData(int amt, Action<byte[]> callback)
+        public void ReadData(long amt, Action<byte[]> callback)
         {
             // allocate the buffer
             byte[] buffer = new byte[amt];
             // start reading from the stream
-            var result = base.GetStream().BeginRead(buffer, 0, amt, _callback, null);
+            var result = base.GetStream().BeginRead(buffer, 0, (int)amt, _callback, null);
 
             // wait until completion
             if (!result.IsCompleted)
